requirements.txt
Flask==1.1.2
Flask_Cors==3.0.10
flask_ngrok==0.0.25
matplotlib==3.5.2
Pillow==9.5.0
tensorflow==2.12.0
tensorflow_intel==2.11.0
Werkzeug==2.0.3
joblib==1.2.0
requests==2.28.1
itsdangerous==2.1.2
